+++ modules/layers/src/index.ts
@@ -22,5 +22,5 @@ export {
 export type {SkyboxLayerProps} from './skybox-layer/skybox-layer';
 export {SkyboxLayer} from './skybox-layer/skybox-layer';
 
-export type {NewHeatLayerProps} from './newheat-layer/newheat-layer';
-export {NewHeatLayer} from './newheat-layer/newheat-layer';
+export type {FlameTrailLayerProps} from './flame-trail-layer/flame-trail-layer';
+export {FlameTrailLayer} from './flame-trail-layer/flame-trail-layer';