+++ website/src/examples-sidebar.js
@@ -43,7 +43,6 @@ const sidebars = {
       label: '@deck.gl-community/layers',
       items: [
         'layers/path-outline-and-markers',
-        'layers/newheat',
         'layers/skybox-map-view',
         'layers/skybox-globe',
         'layers/skybox-first-person'